fix histogram oneline view

+++ DataUtils.h
@@ -275,7 +275,7 @@ public:
         o << "  " << m_underf << " [";
         for(size_t i=0; i<this->BinSize(); ++i) {
             double val = m_bins[i];
-            unsigned int lid = floor(val/max * 6);
+            unsigned int lid = (int)floor(val/max * 6);
             o << lut[lid];
         }
         o << "] " << m_overf << " ";
